+++ lib/src/commands/commands.dart
@@ -2,5 +2,6 @@ export 'build_commands/build_command.dart';
 export 'doctor_command.dart';
 export 'flutter_switch_command.dart';
 export 'install_commands/install_command.dart';
+export 'run_command.dart';
 export 'uninstall_command.dart';
 export 'update_command.dart';